grc: moved all usrp1 and usrp2 stuff out of grc and into gr-usrp*/grc
Moved grc xml files, python wrappers (USRP1 only), probe apps + freedesktop files.

When the gr-usrp and gr-usrp2 directories are removed, grc will not have to change.

Minor change: the freedesktop files are always installed now.
This does not mean that they are installed properly with xdg,
it just means that the runtime can have access to the icons.