grc: prevent scrolling to top of canvas
GRC was scrolling to the top of the canvas when clicking.

This occurs when clicking into the console or search tree then back into the canvas, the view is reset to the top left.

Commented out one line in _handle_mouse_button_press in DrawingArea.py -
makes the problem go away.

Fixes #2660