grc: prevent search keystrokes from modifying flowgraph
A confluence of bugs introduced while updating GRC to Python 3 results 
in keystrokes being sent to the drawing area while typing in the block 
search box.

First, the drawing area needs to be placed in a Viewport before being 
added into a ScrolledWindow. Fixing this allows the drawing area to grab 
focus as intended when it is clicked.

Second, the drawing area needs to unselect all blocks when it loses 
focus so that the relevant keystroke actions can be disabled. This 
functionality was commented out during the transition to Python 3, and 
I've restored it here.