grc: In flowgraph, use closeEvent() to stop flow graph and not aboutToQuit
So originally I introduced the aboutToQuit() stuff in the hope to
cleanly shutdown the flow graph, including the Qt widgets blocks
before the Qt stuff is torn down.

Turns out that in aboutToQuit some stuff is already invalid, especially
in OpenGL and so causes SegFault on exit.

Instead this patch modify the logic so that we stop the flow graph before
Qt even begin disabling things.

The two shutdown path that exits are :
 - The sig_handler in case we get TERM or INT signals
 - The closeEvent() of the main window

This makes sure that both path call tb.stop() tb.wait() and also
any snippet 'main_after_stop'.

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>