The uninstall script did not remove the previously installed freedesktop icons and desktop files. Now grc_setup_freedesktop uninstall is called in cmake_uninstall.cmake
As far as I can see the icon and menu files are only needed in the freedesktop environment.
It's not necessary to put them into gnuradio install path.

So I removed these copies and installed these files directly from the source directory
using xdg-icon-resource and other xdg-utils.

Giving grc_setup_freedesktop execute permissions in the build directory avoids copying it
to the libexec directory. So even if ENABLE_POSTINSTALL=ON no libexec directory will be created.

Signed-off-by: Volker Schroer <3470424+dl1ksv@users.noreply.github.com>