grc: Check flow graph format version
The constant FLOW_GRAPH_FILE_FORMAT_VERSION is now used as a check while
opening flow graphs. If a flow graph is opened which exceeds the
FLOW_GRAPH_FILE_FORMAT_VERSION known to this version of GRC, then the
assumption is that the .grc file was created with a later version of GRC,
and can't be opened in this version of GRC. In this case, the loading is
aborted.

Signed-off-by: Martin Braun <martin.braun@ettus.com>