grc: Don't overwrite port domain on auto-type deduction
Fixes the following issue: Create a port with automatic type deduction
which is *not* in the streaming domain (e.g., create an RFNoC RX
streamer block with no changes to settings from the default).

GRC will attempt to guess the dtype of this port (e.g., if it is
connected to another block that does sc16, it will now also be sc16).
However, due to the order in which objects are evaluated, the first time
this is attempted is before the connection objects are generated, which
means the auto-type deduction will fail the first time it is attempted.

The issue is, on failure, we re-set the port domain to 'stream'
unconditionally. If the port domain was anything else (e.g., as above,
'rfnoc') then it will temporarily disable any domain-specific settings
until the connection object is created, which is too late to import
connection-and-domain-specific settings.

With this change, we simply don't touch the domain when the auto-type
deduction fails.

Signed-off-by: Martin Braun <martin.braun@ettus.com>